Add files via upload
Add flutter dependencies in pubspec.yaml dependencies:
  flutter:
    sdk: flutter
  provider: ^6.0.1
  http: ^0.13.3
Adding Networking & Api Integeration
_isLoading and _errorMessage are used to manage loading state and error messages.
fetchFruits method sets _isLoading to true before making the HTTP request and sets it to false in the finally block after completing the request or encountering an error.
If the HTTP request succeeds (response.statusCode == 200), _fruits are updated with the fetched data. Otherwise, _errorMessage is set and _fruits are set to _fallbackFruits.
If an exception is caught during the HTTP request, _errorMessage is set and _fruits are set to _fallbackFruits.